+++ example_application/plugins/@capacitor/action-sheet.js
@@ -29,7 +29,7 @@ var ActionSheetButtonStyle;
 const ActionSheetOptionStyle = ActionSheetButtonStyle;
 
 const ActionSheet = registerPlugin('ActionSheet', {
-    web: () => import('../common/web-31f5983c.js').then(m => new m.ActionSheetWeb()),
+    web: () => import('../common/web-266235e8.js').then(m => new m.ActionSheetWeb()),
 });
 
 export { ActionSheet, ActionSheetButtonStyle, ActionSheetOptionStyle };