+++ example_application/plugins/@capacitor/app.js
@@ -1,7 +1,7 @@
 import { registerPlugin } from './core.js';
 
 const App = registerPlugin('App', {
-    web: () => import('../common/web-b1d9d3d2.js').then(m => new m.AppWeb()),
+    web: () => import('../common/web-d56acca6.js').then(m => new m.AppWeb()),
 });
 
 export { App };