+++ example_application/plugins/@capacitor/device.js
@@ -1,7 +1,7 @@
 import { registerPlugin } from './core.js';
 
 const Device = registerPlugin('Device', {
-    web: () => import('../common/web-67daad1f.js').then(m => new m.DeviceWeb()),
+    web: () => import('../common/web-47224e2e.js').then(m => new m.DeviceWeb()),
 });
 
 export { Device };