+++ example_application/plugins/@capacitor/dialog.js
@@ -1,7 +1,7 @@
 import { registerPlugin } from './core.js';
 
 const Dialog = registerPlugin('Dialog', {
-    web: () => import('../common/web-fe308abf.js').then(m => new m.DialogWeb()),
+    web: () => import('../common/web-53a6aa72.js').then(m => new m.DialogWeb()),
 });
 
 export { Dialog };