+++ example_application/plugins/@capacitor/network.js
@@ -1,7 +1,7 @@
 import { registerPlugin } from './core.js';
 
 const Network = registerPlugin('Network', {
-    web: () => import('../common/web-08e24156.js').then(m => new m.NetworkWeb()),
+    web: () => import('../common/web-60057aea.js').then(m => new m.NetworkWeb()),
 });
 
 export { Network };