+++ example_application/plugins/@capacitor/preferences.js
@@ -1,7 +1,7 @@
 import { registerPlugin } from './core.js';
 
 const Preferences = registerPlugin('Preferences', {
-    web: () => import('../common/web-672d63c0.js').then(m => new m.PreferencesWeb()),
+    web: () => import('../common/web-7dc896a8.js').then(m => new m.PreferencesWeb()),
 });
 
 export { Preferences };