+++ example_application/plugins/@capacitor/share.js
@@ -1,7 +1,7 @@
 import { registerPlugin } from './core.js';
 
 const Share = registerPlugin('Share', {
-    web: () => import('../common/web-73f790ea.js').then(m => new m.ShareWeb()),
+    web: () => import('../common/web-bb419041.js').then(m => new m.ShareWeb()),
 });
 
 export { Share };