+++ example_application/plugins/@capacitor/toast.js
@@ -1,7 +1,7 @@
 import { registerPlugin } from './core.js';
 
 const Toast = registerPlugin('Toast', {
-    web: () => import('../common/web-c35afa81.js').then(m => new m.ToastWeb()),
+    web: () => import('../common/web-90727218.js').then(m => new m.ToastWeb()),
 });
 
 export { Toast };